backend/requirements.txt
langgraph
langchain
langsmith
python-dotenv
langchain-groq
Ipython
langchain_tavily
openai
google-genai